Introduce zsync metadata creation API
Extracted from the make.c command line utility.
