+++ c/configure.ac
@@ -69,5 +69,6 @@ dnl --- output
 AC_SUBST(ac_aux_dir)
 AC_CONFIG_HEADERS([config.h])
 AC_CONFIG_FILES([Makefile librcksum/Makefile zlib/Makefile libzsync/Makefile doc/Makefile])
+AC_DEFINE(WITH_ZLIB, 1, "Make use of zlib")
 AC_OUTPUT
